Move record components to their own directory.
