+++ scripts/components/record/AddForm.js
