+++ scripts/components/record/BulkForm.js
@@ -2,8 +2,8 @@ import React, { Component } from "react";
 import { Link } from "react-router";
 import Form from "react-jsonschema-form";
 
-import Spinner from "./Spinner";
-import JSONEditor from "./JSONEditor";
+import Spinner from "../Spinner";
+import JSONEditor from "../JSONEditor";
 import {
   extendSchemaWithAttachment,
   extendUiSchemaWithAttachment
