+++ scripts/containers/BulkFormPage.js
@@ -2,7 +2,7 @@ import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 import { push as updatePath } from "react-router-redux";
 
-import BulkForm from "../components/BulkForm";
+import BulkForm from "../components/record/BulkForm";
 import * as CollectionActions from "../actions/collection";
 import * as NotificationsActions from "../actions/notifications";
